Tas 329 fix confirmation popup issues

+++ src/stores/boards.ts
@@ -2,7 +2,8 @@ import type { Board, BoardColumn, FormSubsetItem } from '../api/boardsTypes'
 import type {
   CollectionReference,
   DocumentReference,
-  DocumentData
+  DocumentData,
+  FirestoreError
 } from 'firebase/firestore'
 import { defineStore } from 'pinia'
 import { ref, computed } from 'vue'
@@ -11,6 +12,7 @@ import {
   orderBy,
   collection,
   doc,
+  getDoc,
   getDocs,
   addDoc,
   updateDoc,
@@ -42,17 +44,25 @@ export const useBoardsStore = defineStore('boards', () => {
       orderBy('createdAt', 'desc')
     )
 
-    const boardDocs = await getDocs(boardsColRefOrdered)
+    try {
+      const boardDocs = await getDocs(boardsColRefOrdered)
 
-    if (boardDocs.docs.length !== 0) {
-      boards.value = boardDocs.docs.map((snap) => ({
-        ...(snap.data() as Omit<Board, 'boardID'>),
-        boardID: snap.id
-      }))
-    }
+      if (boardDocs == null) throw new Error()
 
-    if (boardDocs.docs.length === 0) {
-      boards.value = []
+      if (boardDocs.docs.length !== 0) {
+        boards.value = boardDocs.docs.map((snap) => ({
+          ...(snap.data() as Omit<Board, 'boardID'>),
+          boardID: snap.id
+        }))
+      }
+
+      if (boardDocs.docs.length === 0) {
+        boards.value = []
+      }
+
+      return true
+    } catch (err) {
+      return (err as FirestoreError).code
     }
   }
 
@@ -68,21 +78,50 @@ export const useBoardsStore = defineStore('boards', () => {
         orderBy('createdAt', 'asc')
       )
 
-      const columnDocs = (await getDocs(columnsColRefOrdered)).docs
-      boardColumns.value =
-        columnDocs.length !== 0
-          ? await Promise.all(
-              columnDocs.map(async (columnDoc) => {
-                return {
-                  ...(columnDoc.data() as Omit<BoardColumn, 'columnID'>),
-                  columnID: columnDoc.id
-                }
-              })
-            )
-          : []
+      try {
+        const columnDocs = (await getDocs(columnsColRefOrdered)).docs
 
-      await tasksStore.getTasks(columnsColRef, boardColumns.value)
+        if (columnDocs == null) throw new Error()
+
+        boardColumns.value =
+          columnDocs.length !== 0
+            ? await Promise.all(
+                columnDocs.map(async (columnDoc) => {
+                  return {
+                    ...(columnDoc.data() as Omit<BoardColumn, 'columnID'>),
+                    columnID: columnDoc.id
+                  }
+                })
+              )
+            : []
+
+        if (columnDocs.length !== 0) {
+          const response = await tasksStore.getTasks(
+            columnsColRef,
+            boardColumns.value
+          )
+
+          if (response !== true) throw new Error('wrong response')
+        }
+        return true
+      } catch (err) {
+        return (err as FirestoreError).code
+      }
     }
+    return 'wrong response'
+  }
+
+  const saveCurrentBoard = async (newBoard: Board) => {
+    currentBoard.value = newBoard
+    localStorage.setItem(
+      `TM-currentBoard-${userStore.userID}`,
+      JSON.stringify(currentBoard.value)
+    )
+    const response = await getColumns()
+
+    if (response !== true) return response
+
+    return true
   }
 
   const addDocToFirestore = async (
@@ -99,36 +138,53 @@ export const useBoardsStore = defineStore('boards', () => {
     docRef: DocumentReference<DocumentData>,
     name: string
   ) => {
-    return await updateDoc(docRef, {
+    const nameBefore = name
+    await updateDoc(docRef, {
       name
     })
+    const nameAfter = ((await getDoc(docRef)).data() as Board).name
+
+    if (nameBefore !== nameAfter) return 'wrong response'
+
+    return true
   }
 
   const addNewBoard = async (
     boardName: Board['name'],
     boardColumns: string[]
   ) => {
-    const addedDocRef = await addDocToFirestore(
-      boardsColRefGlobal.value as CollectionReference<DocumentData>,
-      boardName
-    )
+    try {
+      const addedDocRef = await addDocToFirestore(
+        boardsColRefGlobal.value as CollectionReference<DocumentData>,
+        boardName
+      )
 
-    if (addedDocRef) {
-      const columnsColRef = collection(db, `${addedDocRef.path}/columns`)
+      if (addedDocRef == null) throw new Error('custom error')
 
-      boardColumns.forEach(async (column) => {
-        await addDocToFirestore(columnsColRef, column)
-      })
+      if (addedDocRef) {
+        const columnsColRef = collection(db, `${addedDocRef.path}/columns`)
+
+        boardColumns.forEach(async (column) => {
+          await addDocToFirestore(columnsColRef, column)
+        })
+      }
+
+      const boardsResponse = await getBoards()
+
+      if (boardsResponse === true) {
+        try {
+          const saveBoardResponse = await saveCurrentBoard(boards.value[0])
+
+          if (saveBoardResponse !== true) throw new Error(saveBoardResponse)
+        } catch (err) {
+          return (err as FirestoreError).code
+        }
+      }
+
+      return true
+    } catch (err) {
+      return (err as FirestoreError).code
     }
-
-    await getBoards()
-
-    currentBoard.value = boards.value[0]
-    localStorage.setItem(
-      `TM-currentBoard-${userStore.userID}`,
-      JSON.stringify(currentBoard.value)
-    )
-    await getColumns()
   }
 
   const editBoard = async (
@@ -151,29 +207,45 @@ export const useBoardsStore = defineStore('boards', () => {
       areColumnsNamesSame
     ].every((item) => item === true)
 
-    if (isFormNotChanged) return
+    if (isFormNotChanged) return true
 
     const docToEditRef = doc(
       boardsColRefGlobal.value as CollectionReference<DocumentData>,
       currentBoardID.value as string
     )
-    const lastCurrentBoardID = (currentBoard.value as Board).boardID
+    const lastCurrentBoardID = currentBoardID.value
 
     if (!isBoardNameSame) {
-      await updateFirestoreDoc(docToEditRef, boardName)
+      try {
+        const response = await updateFirestoreDoc(docToEditRef, boardName)
 
-      if (areColumnsNamesSame) {
-        await getBoards()
+        if (response !== true) throw new Error()
 
-        currentBoard.value =
-          boards.value.find((board) => board.boardID === lastCurrentBoardID) ||
-          boards.value[0]
-        localStorage.setItem(
-          `TM-currentBoard-${userStore.userID}`,
-          JSON.stringify(currentBoard.value)
-        )
-        await getColumns()
-        return
+        if (areColumnsNamesSame) {
+          try {
+            const response = await getBoards()
+
+            if (response !== true) throw new Error(response)
+          } catch (err) {
+            return (err as FirestoreError).code
+          }
+
+          const newBoard =
+            boards.value.find(
+              (board) => board.boardID === lastCurrentBoardID
+            ) || boards.value[0]
+
+          try {
+            const saveBoardResponse = await saveCurrentBoard(newBoard)
+
+            if (saveBoardResponse !== true) throw new Error(saveBoardResponse)
+          } catch (err) {
+            return (err as FirestoreError).code
+          }
+          return true
+        }
+      } catch (err) {
+        return (err as FirestoreError).code || 'wrong response'
       }
     }
 
@@ -191,7 +263,13 @@ export const useBoardsStore = defineStore('boards', () => {
     })
     if (noRespectiveColumns) {
       noRespectiveColumns.forEach(async ({ name }) => {
-        await addDocToFirestore(columnsColRef, name)
+        try {
+          const response = await addDocToFirestore(columnsColRef, name)
+
+          if (response == null) throw new Error()
+        } catch (err) {
+          return (err as FirestoreError).code
+        }
       })
     }
 
@@ -199,32 +277,45 @@ export const useBoardsStore = defineStore('boards', () => {
       boardColumns.value.forEach(async ({ columnID, name }) => {
         const columnDocRef = doc(columnsColRef, columnID)
 
-        const respectiveSubtask = updatedColumns.find(
+        const respectiveColumn = updatedColumns.find(
           ({ id }) => id === columnID
         )
-        const isSubtaskNameSame = respectiveSubtask?.name === name
+        const isColumnNameSame = respectiveColumn?.name === name
 
-        if (respectiveSubtask && isSubtaskNameSame) return
+        if (respectiveColumn && isColumnNameSame) return
 
-        if (respectiveSubtask == null) {
-          await deleteDoc(columnDocRef)
+        if (respectiveColumn == null) {
+          try {
+            await deleteDoc(columnDocRef)
+          } catch (err) {
+            return (err as FirestoreError).code
+          }
           return
         }
 
-        await updateFirestoreDoc(columnDocRef, name)
+        try {
+          await updateFirestoreDoc(columnDocRef, respectiveColumn.name)
+        } catch (err) {
+          return (err as FirestoreError).code
+        }
       })
     }
 
-    await getBoards()
-
-    currentBoard.value =
+    const newBoard =
       boards.value.find((board) => board.boardID === lastCurrentBoardID) ||
       boards.value[0]
-    localStorage.setItem(
-      `TM-currentBoard-${userStore.userID}`,
-      JSON.stringify(currentBoard.value)
-    )
-    await getColumns()
+
+    try {
+      const saveBoardResponse = await saveCurrentBoard(newBoard)
+      const boardsResponse = await getBoards()
+
+      if (saveBoardResponse !== true) throw new Error(saveBoardResponse)
+      if (boardsResponse !== true) throw new Error(boardsResponse)
+    } catch (err) {
+      return (err as FirestoreError).code
+    }
+
+    return true
   }
 
   const deleteBoard = async (boardID: Board['boardID']) => {
@@ -235,30 +326,56 @@ export const useBoardsStore = defineStore('boards', () => {
     const columnsColRefs = collection(db, `${boardDocRef.path}/columns`)
     const columnsDocRefs = (await getDocs(columnsColRefs)).docs
 
-    columnsDocRefs.forEach(async (columnDocRef) => {
-      await deleteColumn(columnDocRef.id)
-    })
+    const columnResponses = await Promise.all(
+      columnsDocRefs.map(async (columnDocRef) => {
+        try {
+          const response = await deleteColumn(columnDocRef.id, true)
 
-    await deleteDoc(boardDocRef)
-    await getBoards()
+          if (response !== true) throw new Error(response)
 
-    if (boards.value.length === 0) {
-      currentBoard.value = null
-      localStorage.removeItem(`TM-currentBoard-${userStore.userID}`)
-      await getBoards()
-      await getColumns()
-      return
-    }
-
-    currentBoard.value = boards.value[0]
-    localStorage.setItem(
-      `TM-currentBoard-${userStore.userID}`,
-      JSON.stringify(currentBoard.value)
+          return true
+        } catch (err) {
+          return (err as FirestoreError).code
+        }
+      })
     )
-    await getColumns()
+
+    if (columnResponses.some((response) => response !== true))
+      return 'wrong response'
+
+    try {
+      await deleteDoc(boardDocRef)
+      const boardsResponse = await getBoards()
+
+      if (boardsResponse !== true) throw new Error(boardsResponse)
+
+      if (boards.value.length === 0) {
+        currentBoard.value = null
+        localStorage.removeItem(`TM-currentBoard-${userStore.userID}`)
+      }
+
+      if (boards.value.length > 0) {
+        try {
+          const saveBoardResponse = await saveCurrentBoard(boards.value[0])
+          const columnsResponse = await getColumns()
+
+          if (saveBoardResponse !== true) throw new Error(saveBoardResponse)
+          if (columnsResponse !== true) throw new Error(columnsResponse)
+        } catch (err) {
+          return (err as FirestoreError).code
+        }
+      }
+
+      return true
+    } catch (err) {
+      return (err as FirestoreError).code
+    }
   }
 
-  const deleteColumn = async (columnID: BoardColumn['columnID']) => {
+  const deleteColumn = async (
+    columnID: BoardColumn['columnID'],
+    omitFetch?: true
+  ) => {
     const columnsColRef = collection(
       db,
       `${
@@ -268,27 +385,44 @@ export const useBoardsStore = defineStore('boards', () => {
     const columnDocRef = doc(columnsColRef, columnID)
 
     const tasksColRefs = collection(db, `${columnDocRef.path}/tasks`)
-    const tasksDocRefs = await getDocs(tasksColRefs)
-    if (tasksDocRefs.docs.length !== 0) {
-      tasksDocRefs.forEach(async (tasksDocRef) => {
-        const subtasksColRef = collection(
-          db,
-          `${tasksColRefs.path}/${tasksDocRef.id}/subtasks`
-        )
 
-        const subtasksDocRefs = await getDocs(subtasksColRef)
-        if (subtasksDocRefs.docs.length !== 0) {
-          subtasksDocRefs.forEach(async (subtasksDocRef) => {
-            await deleteDoc(subtasksDocRef.ref)
-          })
+    try {
+      const tasksDocRefs = await getDocs(tasksColRefs)
+
+      if (tasksDocRefs.docs.length !== 0) {
+        tasksDocRefs.forEach(async (tasksDocRef) => {
+          const subtasksColRef = collection(
+            db,
+            `${tasksColRefs.path}/${tasksDocRef.id}/subtasks`
+          )
+
+          const subtasksDocRefs = await getDocs(subtasksColRef)
+          if (subtasksDocRefs.docs.length !== 0) {
+            subtasksDocRefs.forEach(async (subtasksDocRef) => {
+              await deleteDoc(subtasksDocRef.ref)
+            })
+          }
+
+          await deleteDoc(tasksDocRef.ref)
+        })
+      }
+
+      await deleteDoc(columnDocRef)
+
+      if (omitFetch != null) {
+        try {
+          const response = await getColumns()
+
+          if (response !== true) throw new Error(response)
+        } catch (err) {
+          return (err as FirestoreError).code
         }
+      }
 
-        await deleteDoc(tasksDocRef.ref)
-      })
+      return true
+    } catch (err) {
+      return (err as FirestoreError).code
     }
-
-    await deleteDoc(columnDocRef)
-    await getColumns()
   }
 
   return {
@@ -297,6 +431,7 @@ export const useBoardsStore = defineStore('boards', () => {
     boardColumns,
     getBoards,
     getColumns,
+    saveCurrentBoard,
     addNewBoard,
     editBoard,
     deleteBoard,
